Application part complated

+++ CatalogManagement/src/CatalogManagement.Application/ProductGroups/Commands/DeleteProductGroup/DeleteProductGroupByIdCommandHandler.cs
@@ -1,21 +1,23 @@
 ﻿using CatalogManagement.Application.Common.Repositories;
 using CatalogManagement.Domain.ProductGroupAggregate.Errors;
+using CatalogManagement.Domain.ProductGroupAggregate.Events;
 using CatalogManagement.SharedKernel;
 using MediatR;
 
 namespace CatalogManagement.Application.ProductGroups;
-internal class DeleteProductGroupByIdCommandHandler(IProductGroupRepository productGroupRepository)
+internal sealed class DeleteProductGroupByIdCommandHandler(IProductGroupRepository productGroupRepository)
         : IRequestHandler<DeleteProductGroupByIdCommand, Result<bool>>
 {
     private readonly IProductGroupRepository productGroupRepository = productGroupRepository;
 
     public async Task<Result<bool>> Handle(DeleteProductGroupByIdCommand request, CancellationToken cancellationToken)
     {
-        var deletedResult = await productGroupRepository.DeleteByIdAsync(request.Id, cancellationToken);
-        if (deletedResult)
+        var group = await productGroupRepository.GetByIdAsync(request.Id, cancellationToken);
+        if (group is null)
         {
-            return Result<bool>.Success(deletedResult);
+            return ProductGroupError.NotFoundById;
         }
-        return Result<bool>.Fail(ProductGroupError.NotDeleted);
+        group.AddDomainEvent(new ProductGroupDeletedDomainEvent(group.Id));
+        return await productGroupRepository.DeleteByIdAsync(request.Id, cancellationToken);
     }
 }
